+++ config/environments/test.rb
@@ -41,13 +41,13 @@ Rails.application.configure do
   # Disable request forgery protection in test environment.
   config.action_controller.allow_forgery_protection = false
 
-  # config.action_mailer.perform_caching = false
+  config.action_mailer.perform_caching = false
 
   # Tell Action Mailer not to deliver emails to the real world.
   # The :test delivery method accumulates sent emails in the
   # ActionMailer::Base.deliveries array.
-  # config.action_mailer.delivery_method = :test
-  # config.action_mailer.default_url_options = { host: 'localhost', port: 5000 }
+  config.action_mailer.delivery_method = :test
+  config.action_mailer.default_url_options = { host: 'localhost', port: 5000 }
 
   # Print deprecation notices to the stderr.
   config.active_support.deprecation = :stderr