+++ config/schedule.rb
@@ -1,11 +1,14 @@
 # frozen_string_literal: true
 
+# to view cron rules - visit https://crontab.guru
+
 # Synchronize companies
 every 1.hour do
   runner 'Import::SyncCompaniesJob.perform_later'
 end
 
 # Refresh user achievements
+# At minute 15 on every day-of-week from Monday through Friday
 every '15 * * * 1-5' do
   runner 'Users::RefreshAchievementsJob.perform_later'
 end