+++ spec/services/import/sync_repositories_service_spec.rb
@@ -23,6 +23,7 @@ describe Import::SyncRepositoriesService, type: :service do
   let(:sync_reviews_service) { double(Import::SyncReviewsService) }
   let(:reviews_service) { double }
   let(:generate_insights_service) { double(Insights::GenerateService) }
+  let(:mailer) { double }
 
   before do
     create :access_token, tokenable: company
@@ -34,6 +35,8 @@ describe Import::SyncRepositoriesService, type: :service do
     allow(sync_reviews_service).to receive(:new).and_return(reviews_service)
     allow(reviews_service).to receive(:call)
     allow(generate_insights_service).to receive(:call)
+    allow(Users::NotificationMailer).to receive(:repository_access_error_email).and_return(mailer)
+    allow(mailer).to receive(:deliver_now)
   end
 
   it 'calls services and updates repository', :aggregate_failures do
@@ -44,11 +47,41 @@ describe Import::SyncRepositoriesService, type: :service do
     expect(sync_reviews_service).to have_received(:new).with(pull_request: pull_request1)
     expect(sync_comments_service).not_to have_received(:new).with(pull_request: pull_request2)
     expect(sync_reviews_service).not_to have_received(:new).with(pull_request: pull_request2)
-    expect(generate_insights_service).to have_received(:call).with(insightable: repository)
-    expect(generate_insights_service).to have_received(:call).with(insightable: company)
+    expect(generate_insights_service).not_to have_received(:call)
     expect(repository.reload.synced_at).not_to be_nil
     expect(company.reload.accessable).to be_falsy
-
+    expect(company.not_accessable_ticks).to eq 1
     expect(service_call.success?).to be_truthy
+    expect(Users::NotificationMailer).not_to have_received(:repository_access_error_email)
+  end
+
+  context 'for over not accessable company' do
+    before { company.update!(not_accessable_ticks: described_class::NOT_ACCESSABLE_LIMIT_TICKS - 1) }
+
+    it 'calls insights services and updates repository', :aggregate_failures do
+      service_call
+
+      expect(repository.reload.synced_at).not_to be_nil
+      expect(company.reload.accessable).to be_falsy
+      expect(company.not_accessable_ticks).to eq 10
+      expect(service_call.success?).to be_truthy
+      expect(Users::NotificationMailer).to have_received(:repository_access_error_email).with(id: company.user_id)
+    end
+  end
+
+  context 'for accessable repository' do
+    before { repository.update(accessable: true) }
+
+    it 'calls insights services and updates repository', :aggregate_failures do
+      service_call
+
+      expect(generate_insights_service).to have_received(:call).with(insightable: repository)
+      expect(generate_insights_service).to have_received(:call).with(insightable: company)
+      expect(repository.reload.synced_at).not_to be_nil
+      expect(company.reload.accessable).to be_truthy
+      expect(company.not_accessable_ticks).to eq 0
+      expect(service_call.success?).to be_truthy
+      expect(Users::NotificationMailer).not_to have_received(:repository_access_error_email)
+    end
   end
 end